Create .cc for GenMatching
